Properly check if temporal file exists before attempting removal

+++ ufw-cf.sh
@@ -68,7 +68,9 @@ if [ $cfufw_purge -eq 1 ]; then
     cf_ufw_purge
 fi
 
+[ -e /tmp/cloudflare-ips.txt ] && rm /tmp/cloudflare-ips.txt
 touch /tmp/cloudflare-ips.txt
+
 wget https://www.cloudflare.com/ips-v4 -q -O ->> /tmp/cloudflare-ips.txt
 wget https://www.cloudflare.com/ips-v6 -q -O ->> /tmp/cloudflare-ips.txt
 
@@ -76,7 +78,7 @@ for cfip in `cat /tmp/cloudflare-ips.txt`; do
         cf_ufw_add "$cfip"
 done
 
-rm /tmp/cloudflare-ips.txt
+[ -e /tmp/cloudflare-ips.txt ] && rm /tmp/cloudflare-ips.txt
 
 echo 'Done.'
